+++ src/app/(components)/timesheet/modal/day-additional-modal.jsx
@@ -91,7 +91,6 @@ function AdditionModal({ date, onClose, onTimesheetUpdate }) {
     }
 
     if (time === "00:00") {
-      console.log(time);
       setErrorMessage("Please enter a valid time.");
       return;
     }